Handle corner case in ArtifactStore getLatest, which cascades to MaterialPoller
In our setup, we've noticed that objectMetadata request after doing a getLatest fails with 404. One possible code path that leads to this is getLatest returning Revisions.base(), which is problematic, since, we consider it to be a valid revision and try to do a metadata fetch. Added a fix to handle it with logging enabled.